+++ src/808x.h
@@ -266,7 +266,7 @@ struct CPU8086
     }
 
 
-    u32 prefetch_queue_size = 6;
+    u8 prefetch_queue_size = 6;
     u8 prefetch_queue[6] = {};
     u32 prefetch_address{0xFFFFFFFF};
